split concept module (wip), enum_dispatch
